+++ integrations/storage/v2/shared/abs/configuration.go
@@ -1,0 +1,59 @@
+//
+// DISCLAIMER
+//
+// Copyright 2025 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package abs
+
+import (
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
+
+	pbImplStorageV2Shared "github.com/arangodb/kube-arangodb/integrations/storage/v2/shared"
+	"github.com/arangodb/kube-arangodb/pkg/util/azure"
+)
+
+type Configuration struct {
+	BucketName   string
+	BucketPrefix string
+
+	MaxListKeys *int32
+
+	Client azure.Config
+}
+
+func (c Configuration) New() (pbImplStorageV2Shared.IO, error) {
+	prov, err := c.Client.GetCredentials()
+	if err != nil {
+		return nil, err
+	}
+
+	endpoint, err := c.Client.GetEndpoint()
+	if err != nil {
+		return nil, err
+	}
+
+	client, err := azblob.NewClient(endpoint, prov, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return &ios{
+		config: c,
+		client: client.ServiceClient(),
+	}, nil
+}